pyproject.toml
[project]
name = "trading"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "ddgs>=9.8.0",
    "duckduckgo-search>=8.1.1",
    "ipykernel>=6.30.1",
    "langchain>=0.3.27",
    "langchain-community>=0.3.31",
    "langchain-openai>=0.3.35",
    "langgraph>=0.6.11",
    "pandas>=2.3.2",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "yfinance>=0.2.65",
]

[dependency-groups]
dev = [
    "ipykernel>=6.30.1",
]

[tool.uv]
package = true

[tool.setuptools.packages.find]
where = ["."]

[project.scripts]
agent = "src.main:main"
